factored _tobytes and _totext into utils module, added docstrings.
